requirements.txt
functions-framework==3.*
google-cloud-firestore
google-cloud-storage
beautifulsoup4
requests